+++ src/components/counter/styles.js
@@ -2,8 +2,11 @@ import { StyleSheet } from 'react-native';
 
 export default StyleSheet.create({
   text: {
-    marginLeft: 8,
+    fontSize: 12,
+    lineHeight: 16,
     textAlign: 'right',
     backgroundColor: 'transparent',
+    paddingVertical: 2,
+    marginLeft: 8,
   },
 });